third_party/WebKit/LayoutTests/svg/animations/animate-css-xml-attributeType-expected.txt

SVG 1.1 dynamic animation tests
Tests that XML and CSS attributeTypes can be switched between.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS polygon.animatedPoints.getItem(0).x is 100
PASS polygon.points.getItem(0).x is 100
PASS polygon.animatedPoints.getItem(0).x is 300
PASS polygon.animatedPoints.getItem(0).x is 100
PASS parseFloat(getComputedStyle(polygon).opacity) is 0.8
PASS polygon.animatedPoints.getItem(0).x is 100
PASS parseFloat(getComputedStyle(polygon).opacity) is 1.0
PASS polygon.animatedPoints.getItem(0).x is 200
PASS polygon.points.getItem(0).x is 100
PASS successfullyParsed is true

TEST COMPLETE
